+++ yagr-react.d.ts
@@ -1,0 +1,2 @@
+export * from './dist/react';
+export {default} from './dist/react';